minor refactor of migration code
