+++ migrate/cmd.go
@@ -3,22 +3,21 @@ package migrate
 import (
 	"fmt"
 
-	"github.com/kava-labs/kava/migrate/v0_14"
-	"github.com/spf13/cobra"
-
 	"github.com/cosmos/cosmos-sdk/codec"
 	"github.com/cosmos/cosmos-sdk/server"
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	"github.com/cosmos/cosmos-sdk/version"
-
+	"github.com/spf13/cobra"
 	tmtypes "github.com/tendermint/tendermint/types"
+
+	"github.com/kava-labs/kava/migrate/v0_15"
 )
 
 // MigrateGenesisCmd returns a command to execute genesis state migration.
 func MigrateGenesisCmd(_ *server.Context, cdc *codec.Codec) *cobra.Command {
 	cmd := &cobra.Command{
 		Use:     "migrate [genesis-file]",
-		Short:   "Migrate genesis file from kava v0.11 (or v0.12) to v0.14",
+		Short:   "Migrate genesis file from kava v0.14 to v0.15",
 		Long:    "Migrate the source genesis into the current version, sorts it, and print to STDOUT.",
 		Example: fmt.Sprintf(`%s migrate /path/to/genesis.json`, version.ServerName),
 		Args:    cobra.ExactArgs(1),
@@ -30,7 +29,7 @@ func MigrateGenesisCmd(_ *server.Context, cdc *codec.Codec) *cobra.Command {
 				return fmt.Errorf("failed to read genesis document from file %s: %w", importGenesis, err)
 			}
 
-			newGenDoc := v0_14.Migrate(*genDoc)
+			newGenDoc := v0_15.Migrate(*genDoc)
 
 			bz, err := cdc.MarshalJSONIndent(newGenDoc, "", "  ")
 			if err != nil {
